+++ Runtime/Scripts/Export/ImageExport.cs
@@ -58,7 +58,11 @@ namespace GLTFast.Export {
                     return $"{nameWithoutExtension}.{fileExtension}";
                 }
 #endif
-                return $"{m_Texture.name}.{fileExtension}";
+                var name = m_Texture.name;
+                if (string.IsNullOrEmpty(name)) {
+                    name = "texture";
+                }
+                return $"{name}.{fileExtension}";
             }
         }